REST규칙에 맞춰서 수정
